fix: fix Android textures bug
When the game is closed using the back button on Android, the activity containing it is disposed. The runtime VM and all static variables are not disposed though, so lingering references remained in the SpriteManager cache to textures that didn't exist anymore. The SpriteManager cache is now cleared when the game's activity is destroyed.

Also added some further code to properly dispose of gestalt-managed textures. The reason why textures appeared to be randomly swapped in the bug is because the underlying sprites were referencing invalid OpenGL handles that had not been cleaned-up properly. By correctly disposing of these textures upon activity destruction, the invalid handles now show black rects instead.